ev-rec-bot-service/requirements.txt
gunicorn==20.1.0
asyncpg==0.28.0
cloud-sql-python-connector==1.4.1
Flask==2.1.0
langchain==0.0.196
requests==2.31.0
werkzeug==2.3.6
google-cloud-aiplatform==1.33.1
asgiref==3.7.2